Update spec.md
I think the unencoded parameters for each URL should be listed in their own right, and encoded within the URLs -- partly but not only because I cannot tell what should be encoded and what should be as shown, which is why I'm not putting my entire suggestion in directly.  I have added evil extra spaces to each line that is likely to be impacted, so they show up as "suggestible" in the Github web interface.